% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsmr.R
\name{heidi_outlier}
\alias{heidi_outlier}
\title{HEIDI-outlier analysis}
\usage{
heidi_outlier(bzx, bzx_se, bzx_pval, bzy, bzy_se, ldrho, snpid, n_ref, gwas_thresh=5e-8, heidi_outlier_thresh=0.01, nsnps_thresh=10, ld_fdr_thresh=0.05)
}
\arguments{
\item{bzx}{vector, SNP effects on risk factor}

\item{bzx_se}{vector, standard errors of bzx}

\item{bzx_pval}{vector, p values for bzx}

\item{bzy}{vector, SNP effects on disease}

\item{bzy_se}{vector, standard errors of bzy}

\item{ldrho}{LD correlation matrix of the SNPs}

\item{snpid}{genetic instruments}

\item{n_ref}{sample size of the reference sample}

\item{gwas_thresh}{threshold p-value to select instruments from GWAS for risk factor}

\item{heidi_outlier_thresh}{threshold p-value to remove pleiotropic outliers (the default value is 0.01)}

\item{nsnps_thresh}{the minimum number of instruments required for the GSMR analysis (we do not recommend users to set this number smaller than 10)}

\item{ld_r2_thresh}{LD r2 threshold to remove SNPs in high LD}

\item{ld_fdr_thresh}{FDR threshold to remove the chance correlations between SNP instruments}
}
\value{
Retained index of genetic instruments, SNPs with missing values, with non-significant p-values and those in LD.
}
\description{
An analysis to detect and eliminate from the analysis instruments that show significant pleiotropic effects on both risk factor and disease
}
\examples{
data("gsmr")
filtered_index = heidi_outlier(gsmr_data$bzx, gsmr_data$bzx_se, gsmr_data$bzx_pval, gsmr_data$bzy, gsmr_data$bzy_se, ldrho, gsmr_data$SNP, n_ref, 5e-8, 0.01, 10, 0.1, 0.05)

}
